+++ cmd/tempo-cli/main.go
@@ -15,6 +15,7 @@ import (
 	"gopkg.in/yaml.v2"
 
 	"github.com/alecthomas/kong"
+	"github.com/grafana/tempo/tempodb/backend/azure"
 	"github.com/grafana/tempo/tempodb/backend/gcs"
 	"github.com/grafana/tempo/tempodb/backend/s3"
 )
@@ -24,7 +25,7 @@ type globalOptions struct {
 }
 
 type backendOptions struct {
-	Backend string `help:"backend to connect to (s3/gcs/local), optional, overrides backend in config file" enum:",s3,gcs,local"`
+	Backend string `help:"backend to connect to (s3/gcs/local/azure), optional, overrides backend in config file" enum:",s3,gcs,local,azure"`
 	Bucket  string `help:"bucket to scan, optional, overrides bucket in config file"`
 
 	S3Endpoint string `name:"s3-endpoint" help:"s3 endpoint (s3.dualstack.us-east-2.amazonaws.com), optional, overrides endpoint in config file"`
@@ -81,6 +82,7 @@ func loadBackend(b *backendOptions, g *globalOptions) (tempodb_backend.Reader, t
 		cfg.StorageConfig.Trace.Local.Path = b.Bucket
 		cfg.StorageConfig.Trace.GCS.BucketName = b.Bucket
 		cfg.StorageConfig.Trace.S3.Bucket = b.Bucket
+		cfg.StorageConfig.Trace.Azure.ContainerName = b.Bucket
 	}
 
 	if b.S3Endpoint != "" {
@@ -98,6 +100,8 @@ func loadBackend(b *backendOptions, g *globalOptions) (tempodb_backend.Reader, t
 		r, _, c, err = gcs.New(cfg.StorageConfig.Trace.GCS)
 	case "s3":
 		r, _, c, err = s3.New(cfg.StorageConfig.Trace.S3)
+	case "azure":
+		r, _, c, err = azure.New(cfg.StorageConfig.Trace.Azure)
 	default:
 		err = fmt.Errorf("unknown backend %s", cfg.StorageConfig.Trace.Backend)
 	}